Add a registerDial function to mirror the functionality of the mysql driver
This adds a function, registerDial that mirrors the functionality used
in github.com/VividCortex/mysql to set a custom dialer to use for
connections to the database. We can use this to set read/write timeouts
on database connections made via TCP.

The implementation is pretty hacky at the moment, I essentially
transplanted the code used in the mysql driver over to here for
starters.