Refactor custom dialer injection
This reverts the previous implementation in favor of overriding the
defaultDialer’s Dial() methods directly. We check in defaultDialer.Dial
whether a custom dialer has been registered and use it instead if we
find one. This way a custom dialer passed to DialOpen is preserved when
a custom dialer is registered.